Added database classes and data properties for user authorization
Some database modification was necessary in order to support authorization / authentication because we did not want to rely on the integrated database user management.

The main modification is the addition of a "User" class which represents all users. Users are expected to have a unique username data property (the uniqueness constraint is a todo), a passwordHash data property, a userRole object property, and a creation data property. One test user is already preinserted into the database, and additional users should be added under the "namespace" "sdg.user."